# sql: decrease vectorize_row_count_threshold to 0