--- a/pkg/sql/colexec/constants.go
+++ b/pkg/sql/colexec/constants.go
@@ -12,12 +12,9 @@ package colexec
 
 // DefaultVectorizeRowCountThreshold denotes the default row count threshold.
 // When it is met, the vectorized execution engine will be used if possible.
-// The current number 1000 was chosen upon comparing `SELECT count(*) FROM t`
-// query running through the row and the vectorized execution engines on a
-// single node with tables having different number of columns.
-// Note: if you are updating this field, please make sure to update
-// vectorize_threshold logic test accordingly.
-const DefaultVectorizeRowCountThreshold = 1000
+// TODO(yuzefovich): remove this together with vectorize_row_count_threshold
+// setting.
+const DefaultVectorizeRowCountThreshold = 0
 
 // VecMaxOpenFDsLimit specifies the maximum number of open file descriptors
 // that the vectorized engine can have (globally) for use of the temporary