@yuzefovich 2020-10-22
The benchmarks.

----
3 node roachprod cluster
----
1. TPCC, 100 warehouses, 1 min ramp, 5 min duration: **-0.748%** in tpmC
  - old
```
_elapsed_______tpmC____efc__avg(ms)__p50(ms)__p90(ms)__p95(ms)__p99(ms)_pMax(ms)
  300.0s     1203.2  93.6%     28.3     26.2     37.7     46.1     71.3    151.0
```
  - new
```
_elapsed_______tpmC____efc__avg(ms)__p50(ms)__p90(ms)__p95(ms)__p99(ms)_pMax(ms)
  300.0s     1194.2  92.9%     29.9     28.3     39.8     48.2     65.0    117.4
```
2. KV95, 1 min ramp, 5 min duration: **-6.45%** in throughput
  - old
```
Highest sequence written: 126430. Can be passed as --write-seq=R126430 to the next run.

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0        2142885         7143.0      1.0      1.0      1.8      3.0    159.4  read

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0         113060          376.9      2.5      2.4      4.1      7.3     30.4  write

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
  300.0s        0        2255945         7519.8      1.1      1.0      2.4      3.5    159.4  
```
- new
```
Highest sequence written: 118557. Can be passed as --write-seq=R118557 to the next run.

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0        2004183         6680.6      1.1      1.0      2.0      3.5    151.0  read

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
  300.0s        0         106178          353.9      2.6      2.5      4.1      6.6     23.1  write

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
  300.0s        0        2110361         7034.5      1.1      1.0      2.5      3.9    151.0  
```
3. KV95, 5 min ramp, 30 min duration: **-8.87%** in throughput
- old
```
Highest sequence written: 843826. Can be passed as --write-seq=R843826 to the next run.

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 1800.0s        0       14648469         8138.0      0.9      0.9      1.6      2.4     50.3  read

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 1800.0s        0         771129          428.4      2.1      2.0      3.1      4.7     56.6  write

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
 1800.0s        0       15419598         8566.4      0.9      0.9      1.9      2.8     56.6 
```
- new
```
Highest sequence written: 774506. Can be passed as --write-seq=R774506 to the next run.

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 1800.0s        0       13347899         7415.5      1.0      0.9      1.7      2.9    142.6  read

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
 1800.0s        0         703820          391.0      2.3      2.2      3.7      5.8     31.5  write

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
 1800.0s        0       14051719         7806.5      1.0      1.0      2.1      3.3    142.6 
```
----

I trust these numbers a lot more than the ones I got [here](https://github.com/cockroachdb/cockroach/pull/52453#issuecomment-670306184) because benchmarking on the laptop can have significant variance in performance. My current guess is that this difference is mostly caused by the fact that we're pooling `tableReader` objects as well as reusing some of the slices in `tableInfo` objects in the row-by-row engine. I looked briefly into how we could do the same in the vectorized engine, and especially I wanted to remove a lot of duplication between `cFetcher` and `row.fetcher` (as well as `cTableInfo` and `tableInfo`), but it's not that easy.