@jordanlewis 2020-10-22
Thanks for running these benchmarks. What setup did you use? I guess we're not quite ready to merge this yet.