@jordanlewis 2020-10-22
Yes, I meant hardware setup. Thanks!