@awoods187 2020-10-22
On tpcc we typically use a fourth node for the workload to simulate an
application being separate from the database resources. May be worth
adjusting the testing setup in the future.

On Wed, Oct 21, 2020 at 11:30 PM Jordan Lewis <notifications@github.com>
wrote:

> Yes, I meant hardware setup. Thanks!
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/cockroachdb/cockroach/pull/55713#issuecomment-714198363>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AFJ7F75CGTQ32LCAJ6BVEMLSL6ROZANCNFSM4SWWJKFQ>
> .
>
-- 
Andy Woods
Senior Product Manager, Cockroach Labs
andy@cockroachlabs.com