@yuzefovich 2020-10-23
One pretty obvious thing that I missed that is likely to explain the majority of the difference is that without `vectorize_row_count_threshold` we are now always running `SupportsVectorized` check (previously, we would do so only if the threshold is met). That check especially shows up on KV-like workload because we're essentially instantiating the whole flow twice with only the second time running it, and in particular that means that we're creating a `ColBatchScan` and initializing `cFetcher` twice too - `cFetcher.Init` is pretty heavy.

So I think apart from the pooling of objects we'll need to do something about this check. I feel like the most viable option is to make it a lot more lightweight: instead of creating the whole fake flow with all of the components, we could probably just inspect the processor specs and see whether there are any that we refuse to wrap, and if we either support "natively" or can wrap all of the cores, then we'll assume that we can successfully setup the whole flow. I think when `SupportsVectorized` check was introduced, we had a lot more edge cases that needed to be handled, so we went with the fake flow setup, but now that might be unnecessary anymore. It does feel a bit scary to remove that check though.