@yuzefovich 2020-10-23
Hm, to my surprise a prototype with some pooling and a more lightweight support check (#55883) shows big drop in performance between `vectorize_row_count_threshold=1000` and `vectorize_row_count_threshold=0`. I guess we'll have to abandon this PR for now :/