@yuzefovich 2020-10-26
A newer prototype (that is almost production ready) shows an [improvement](https://github.com/cockroachdb/cockroach/pull/55948#issuecomment-716083162) - the performance hit on KV95 if we remove `SupportsVectorized` check in favor of a more light-weight one decreases from 6-8% to 3-4%.