@awoods187 2020-10-28
I'm not sure a 3-4% hit is acceptable for our most important workloads. Have we engaged with KV and other stakeholders on this change? cc @nvanbenschoten