@jordanlewis 2020-10-28
We've planned to merge this early in the cycle, communicate out to the team, and evaluate by the end of the cycle whether or not we can keep this setting as is. Certainly, we are not interested in merging something like this without approval from stakeholders.