@yuzefovich 2020-12-02
TPCC100 numbers again show no significant difference:
- 1 min ramp, 5 min duration
  - GCE:
    - old: `300.0s     1193.2  92.8%     24.1     23.1     30.4     37.7     54.5     92.3`
    - new: `300.0s     1194.8  92.9%     23.9     23.1     30.4     37.7     52.4     96.5`
  - AWS:
    - old: `300.0s     1204.8  93.7%     16.4     15.7     19.9     26.2     39.8    104.9`
    - new: `300.0s     1198.6  93.2%     17.3     16.3     22.0     27.3     41.9    226.5`
- 3 min ramp, 15 min duration
  - GCE:
    - old: `900.0s     1233.3  95.9%     21.9     22.0     26.2     30.4     46.1    121.6`
    - new: `900.0s     1228.7  95.5%     21.9     21.0     27.3     31.5     48.2     88.1`
  - AWS:
    - old: `900.0s     1237.5  96.2%     15.9     15.2     18.9     22.0     35.7    243.3`
    - new: `900.0s     1235.9  96.1%     15.1     14.7     17.8     21.0     33.6    100.7`
- 10 min ramp, 60 min duration
  - GCE:
    - old: `3600.0s     1251.7  97.3%     23.1     23.1     27.3     31.5     48.2    130.0`
    - new: `3600.0s     1252.0  97.4%     22.5     22.0     26.2     30.4     48.2    151.0`
  - AWS:
    - old: `3600.0s     1255.1  97.6%     15.3     14.7     17.8     19.9     33.6    436.2`
    - new: `3600.0s     1253.5  97.5%     15.7     15.2     17.8     19.9     33.6    226.5`

The only interesting detail is that during 60 min run with *old* config on AWS apparently one node died and "new order" transactions were returning an error. I didn't notice it before wiping the cluster, so I'm not sure what happened.

----
And here are the numbers of `movr` workload (with these parameters `--num-histories 10000 --num-promo-codes 10000 --num-rides 5000     --num-users 500 --num-vehicles 150`):
**TODO: this part is incomplete**
- 1 min ramp, 5 min duration
  - GCE:
    - old: `300.0s        0          88306          294.4      3.4      1.2     12.1     44.0     96.5`
    - new: `300.0s        0          86214          287.4      3.5      1.4     11.5     41.9     60.8`
    - **-2.37%** in throughput
  - AWS:
    - old: ``
    - new: ``
    - **** in throughput
- 3 min ramp, 15 min duration
  - GCE:
    - old: `900.0s        0         260376          289.3      3.5      1.4     11.5     41.9     67.1`
    - new: `900.0s        0         254317          282.6      3.5      1.4     13.1     41.9     60.8`
    - **-2.33%** in throughput
  - AWS:
    - old: ``
    - new: ``
    - **** in throughput
- 10 min ramp, 60 min duration
  - GCE:
    - old: `3600.0s        0        1077971          299.4      3.3      1.6     10.0     37.7    209.7`
    - new: ``
    - **** in throughput
  - AWS:
    - old: `3600.0s        0        1488654          413.5      2.4      1.4      6.8     22.0    436.2`
    - new: ``
    - **** in throughput